fix: refactor handlers
